assert that exploratory Bringer owns no cards before deleting
This is the one and only time a spirit is deleted in the forward
direction of a migration.

This allows us to replace the old output message of "0 cards using
Exploratory Bringer", which requires a human to interpret and understand
this message, with the "silence is golden" principle: with 0 cards
owned, say nothing.

+++ pbf/migrations/0053_exploratory_bringer.py
@@ -20,9 +20,12 @@ def convert_exploratory_bringer(apps, schema_editor):
 def delete_exploratory_bringer(apps, schema_editor):
     GamePlayer = apps.get_model('pbf', 'GamePlayer')
     Spirit = apps.get_model('pbf', 'Spirit')
+    Card = apps.get_model('pbf', 'Card')
 
     if GamePlayer.objects.filter(spirit__name='Exploratory Bringer').exists():
         raise Exception("Someone is using Exploratory Bringer")
+    if cards := Card.objects.filter(spirit__name='Exploratory Bringer').values_list('name', flat=True):
+        raise Exception(f"Exploratory Bringer should not own any cards: {list(cards)}")
 
     try:
         Spirit.objects.get(name='Exploratory Bringer').delete()
